Polymer Predictor/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
rdkit-pypi==2022.9.5
matplotlib==3.8.4
